+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # CHANGE LOG
+
+## [Pre-release]
+- Adds Cucumber & Playwright test setup
+
 ## [0.14.0]
 - Upgrade to NYPL Design System version 1.0.0
 - Decouples search query into separate states for each individual form field
@@ -8,7 +12,6 @@
 - Upgrade to Web Reader version 3.1.0
 - Adds author and work title to edition detail page
 - Adds feature flag functionality
-- Adds Cucumber & Playwright test setup
 
 ## [0.13.1]
 - Added "Back to search results" link for works and editions