+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Pre-release]
 - Adds Cucumber & Playwright test setup
+- Fix QA deploy
 
 ## [0.14.0]
 - Upgrade to NYPL Design System version 1.0.0